#5 P/Cパターンに分離
