+++ app/features/task/components/TaskList/index.tsx
@@ -1,21 +1,14 @@
 import { Box, List } from '@mantine/core';
+import { FC } from 'react';
 
-export const TaskList = () => {
-	const tasks = [
-		{
-			id: 1,
-			title: 'Task 1',
-		},
-		{
-			id: 2,
-			title: 'Task 2',
-		},
-		{
-			id: 3,
-			title: 'Task 3',
-		},
-	];
+type Props = {
+	tasks: {
+		id: number;
+		title: string;
+	}[];
+};
 
+export const TaskList: FC<Props> = ({ tasks }) => {
 	return (
 		<Box>
 			<List>
