+++ app/features/task/components/TaskListContainer/index.tsx
@@ -1,0 +1,21 @@
+import { FC } from "react";
+import { TaskList } from "../TaskList";
+
+export const TaskListContainer: FC = () => {
+  const tasks = [
+		{
+			id: 1,
+			title: 'Task 1',
+		},
+		{
+			id: 2,
+			title: 'Task 2',
+		},
+		{
+			id: 3,
+			title: 'Task 3',
+		},
+	];
+
+  return <TaskList tasks={tasks} />;
+}
